src/main/java/com/example/carserviceapp/dto/mapper/ProductMapper.java
package com.example.carserviceapp.dto.mapper;

import com.example.carserviceapp.dto.request.ProductRequestDto;
import com.example.carserviceapp.dto.response.ProductResponseDto;
import com.example.carserviceapp.model.Product;
import org.springframework.stereotype.Component;

@Component
public class ProductMapper implements ResponseDtoMapper<ProductResponseDto, Product>,
        RequestDtoMapper<ProductRequestDto, Product> {

    @Override
    public Product mapToModel(ProductRequestDto dto) {
        Product product = new Product();
        product.setPrice(dto.getPrice());
        product.setTitle(dto.getTitle());
        return product;
    }

    @Override
    public ProductResponseDto mapToDto(Product product) {
        ProductResponseDto dto = new ProductResponseDto();
        dto.setId(product.getId());
        dto.setPrice(product.getPrice());
        dto.setTitle(product.getTitle());
        return dto;
    }
}
